Move to C++20 (#113)
* Move to C++20

* std::enable_if to requires

* Updated build tools

* clang compilation fix

* Update build tools (for SDL compilation)

* GCC compilation fixes

* clang 18 doesn't has zoned_time, return workaround

* Mac compilation fix

* Backward update

* Missed FO_STACK_TRACE_ENTRY

* iOS sdk 13 -> 18

* AnyData float formatting fix

* Exclude libunwind for mac compilation fix

* Formatting fix

* WinApi includes fix

* Missed include

* Use std::ranges

* Build fixes

* Build fixes

* Compilation fixes

* uint -> uint32

* int -> int32

* float -> float32, double -> float64

* Arithmetic conversion from static_cast to numeric_cast

* Simplifying of int/uint usage, move mostly to int

* GCC compilation fix

* Renaming

* Common module split to smaller Essentials modules

* Compilation fix

* Compilation fix

* Refactroing

* Compilation fix

* Compilation fix

* lerp fix

+++ Source/Applications/ASCompilerApp.cpp
@@ -34,9 +34,7 @@
 #include "Common.h"
 
 #include "Application.h"
-#include "DiskFileSystem.h"
 #include "FileSystem.h"
-#include "Log.h"
 #include "ScriptSystem.h"
 #include "Settings.h"
 
@@ -59,7 +57,7 @@ int main(int argc, char** argv)
     FO_STACK_TRACE_ENTRY();
 
     try {
-        InitApp(argc, argv, AppInitFlags::DisableLogTags);
+        InitApp(numeric_cast<int32>(argc), argv, AppInitFlags::DisableLogTags);
 
         FO_RUNTIME_ASSERT(!App->Settings.BakeOutput.empty());
